requirements.txt
fastapi[all]
python-dotenv
openai
motor
